tester evoluted, and it's beautiful now, let's use it always and will help very much, but now let's focus on time, and test more and more

+++ app/philo/3_utils_1.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils_1.c                                          :+:      :+:    :+:   */
+/*   3_utils_1.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: argel <argel@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/15 18:41:05 by acapela-          #+#    #+#             */
-/*   Updated: 2022/06/22 10:42:00 by argel            ###   ########.fr       */
+/*   Updated: 2022/06/23 12:05:19 by argel            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,9 +61,9 @@ int	ft_putstr_fd(char *s, int fd)
 	return (bytes_read);
 }
 
-void	ft_putnbr_fd(int n, int fd)
+void	ft_putnbr_fd(long int n, int fd)
 {
-	int		mais1;
+	long int		mais1;
 	char	c;
 
 	mais1 = 0;
